querySelectorAll not finding proper links
Error when not being able to load cards via the `a` tag selector. The only `a` tag it finds in the troubleshooting guide. In the case of it not being able to load the resources, the selector returns null. Checking getAttribute on an undefined variable throws an error.

`Your browser was unable to load all of Trello's resources. They may have been blocked by your firewall, proxy or browser configuration.`